add PostListComponent
